domain classes refactor
