+++ test/Domain.UnitTests/CustomerTests/CustomerTest.cs
@@ -1,6 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
 using Bogus;
+using Domain.Entities;
 using Domain.UnitTests._Builders;
 using Domain.UnitTests._Common;
 using ExpectedObjects;
@@ -57,29 +57,5 @@ namespace Domain.UnitTests.CustomerTests
         {
             Assert.Throws<ArgumentException>(() => CustomerBuilder.New().WithName(name).Build()).WithMessage(ExceptionMessage.DOMAIN_CUSTOMER_NAME_INVALID);
         }
-    }
-
-
-
-    public class Customer
-    {
-        public Customer(string name, string email)
-        {
-            if (string.IsNullOrWhiteSpace(email))
-                throw new ArgumentException(ExceptionMessage.DOMAIN_CUSTOMER_EMAIL_INVALID);
-
-            if (string.IsNullOrWhiteSpace(name))
-                throw new ArgumentException(ExceptionMessage.DOMAIN_CUSTOMER_NAME_INVALID);
-
-            Name = name;
-            Email = email;
-            Orders = new List<Order>();
-        }
-
-
-        public string Name { get; }
-        public string Email { get; }
-
-        public List<Order> Orders { get; private set; } //todo test this?
     }
 }
